+++ src/commands/decompress.rs
@@ -5,6 +5,7 @@ use std::{
 };
 
 use fs_err as fs;
+use crate::utils::landlock;
 
 #[cfg(not(feature = "bzip3"))]
 use crate::archive;
@@ -316,7 +317,21 @@ fn execute_decompression(
     is_output_dir_provided: bool,
     is_smart_unpack: bool,
 ) -> crate::Result<ControlFlow<(), usize>> {
-    if is_smart_unpack {
+
+    // init landlock sandbox to restrict file system write access to output_dir
+    // The output directory iseither specified with the -d option or the current working directory is used
+    // TODO: restrict acess to the current working directory to allow only creating new files
+   
+    // TODO: move to unpack and smart_unpack to cover the differetn dirctories used for
+    // decompression
+
+    //if !input_is_stdin && options.remove {
+        //permit write access to input_file_path
+    //} else {
+        landlock::init_sandbox(&[output_dir]);
+    //}
+    
+     if is_smart_unpack {
         return smart_unpack(unpack_fn, output_dir, output_file_path, question_policy);
     }
 
@@ -378,6 +393,9 @@ fn smart_unpack(
         "Created temporary directory {} to hold decompressed elements",
         nice_directory_display(temp_dir_path)
     ));
+
+    //first attempt to restict to the tmp file and allow only to rename it in the parent 
+    //landlock::init_sandbox(Some(temp_dir_path));
 
     let files = unpack_fn(temp_dir_path)?;